Editable mindmap title
You can now click on the title of a mindmap and edit it. This does mean
that you CAN’T click on the title and DRAG the mindmap anymore, but you
can still drag it by any other part of the header div. Clicking “Save”
will only create a new node if this map has never been saved before -
if it has been saved, clicking “Save” will overwrite it. If you edit
the name and then click “Save”, no new node will be made but the name
of the existing node will be changed. Clicking “Save As” creates a new
copy of the map, with whatever new name you choose to give it.